Machine Learning + API Databases/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
